+++ src/main/java/agents/flexibility/dynamicProgramming/State.java
@@ -1,0 +1,5 @@
+package agents.flexibility.dynamicProgramming;
+
+public interface State {
+	
+}